+++ packages/compiler/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @docvia/compiler
 
+## 0.2.1
+
+### Patch Changes
+
+- Updated dependencies [6adfee1]
+  - @docvia/runtime@0.3.0
+
 ## 0.2.0
 
 ### Minor Changes